examples/requirements.txt
azure-monitor-opentelemetry-exporter==1.0.0b24